+ login/register progress

+++ src/main/java/com/kgaciarz/bookyourtable/model/Role.java
@@ -4,30 +4,29 @@ package com.kgaciarz.bookyourtable.model;
 import javax.persistence.*;
 
 @Entity
-@Table (name = "roles")
+@Table(name = "roles")
 public class Role {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
 
     @Enumerated(EnumType.STRING)
     @Column(length = 20)
     private ERole name;
 
-    public Role(){
+    public Role() {
 
     }
 
-    public Role(ERole name){
+    public Role(ERole name) {
         this.name = name;
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
